fix(RecycleScroller): Introduce an item wrapper to reduce re-render
When rendering a slot, Vue will act as if the slot content was expanded
in-line at the <slot>'s place.

Having the slots in a v-for therefore can trigger too much rerender for the
slot content, and was generally bad for performance.

Wrapping them in a component should provide some performance improvements as
well as solving the hover update overhead.